Import netevent version 2
This is actually a completely new write up of an equivalent
and improved feature set.

The previous code is in the stable-1 branch.

In part this is me preparing for a bigger multi-box plus
multi-VM setup, as well as me getting rid of my ugly shell
scripts to maintain socat & netevent processes for each
shared device.

Another reason for doing this was the fact that at some
point I realized there were actually quite a handful of
stars on the netevent's github repo and the fact that the
old readme was really annoying to read & understand.
So, enjoy.
Suggestions are always welcome, just open issues on github.

Signed-off-by: Wolfgang Bumiller <wry.git@bumiller.com>