create: add --listen / --on-close parameters
This allows creating a 'netevent create' process listening
on a socket rather than on stdin in order to keep its
created event devices around. This allows restarting the
source daemon and reconnecting it without losing the device
nodes in /dev/input.
For example, if the input event devices are being passed to
a qemu process you can now restart the source daemon without
breaking the VM input devices.

Signed-off-by: Wolfgang Bumiller <wry.git@bumiller.com>